Fixed issue in PmAgentController::monitorProcess. (#75)
If we try to debug cmpackagemanager in Xcode cmpackagemanager produced ECHILD signal and waiting inside monitorProcess
in waitpid stopped. This led to the killing of the debugging process inside monitorProcess. This did not allow us
to debug this process. Got rid of above described killing.
Just continued monitor loop which lead to second call to waitpid on the same process.